refactor: put all navigation config in one place and rename
