+++ src/components/header/components/Navigation.jsx
@@ -7,8 +7,7 @@ import { queryAtom } from '@/config/searchboxConfig';
 import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
 
 // Import Config for the header
-import { categoryPageFilterAttribute } from '@/config/categoryConfig';
-import { linksHeader, selectorNavigationRef } from '@/config/headerConfig';
+import { categoryPageFilterAttribute, linksHeader, selectorNavigationRef } from '@/config/navigationConfig';
 
 // Import Recoil config
 import {
@@ -63,11 +62,10 @@ const Navigation = ({ isMenuOpen, setIsMenuOpen, mobile, tablet }) => {
 
   return (
     <ul
-      className={`${
-        isMenuOpen
-          ? 'container-mobile__navList-items'
-          : 'container__header-nav__links'
-      } `}
+      className={`${isMenuOpen
+        ? 'container-mobile__navList-items'
+        : 'container__header-nav__links'
+        } `}
     >
       {links.map((link, i) => (
         <li
@@ -104,7 +102,7 @@ const Navigation = ({ isMenuOpen, setIsMenuOpen, mobile, tablet }) => {
           <p
             className={
               highlightingCat() === link.name.toLowerCase() ||
-              state?.name === link.name
+                state?.name === link.name
                 ? 'selected'
                 : ''
             }
