--- src/config/categoryConfig.js
@@ -1,10 +1,0 @@
-// ------------------------------------------
-// Category Pages
-// ------------------------------------------
-
-// ------------------------------------------
-// This exported attribute name is used to generate results for Category pages
-// It will create a filter which will be used when a category page link is clicked on
-// Please refer to linksHeader in headerConfig to define the category page values
-// ------------------------------------------
-export const categoryPageFilterAttribute = 'hierarchicalCategories.lvl0';
